moved ArticlePreview to ArticleList
