+++ src/components/common/ArticleList/ArticlePreview/ArticlePreview.jsx
@@ -1,16 +1,16 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
-import Tag from '../Tag/Tag';
-import TagsRow from '../TagsRow/TagsRow';
-import { LikeDefaultIcon } from '../../../images/icons';
+import Tag from '../../Tag/Tag';
+import TagsRow from '../../TagsRow/TagsRow';
+import { LikeDefaultIcon } from '../../../../images/icons';
 
 import {
   ARTICLE_FAVORITED,
   ARTICLE_UNFAVORITED,
-} from '../../../constants/actionTypes';
+} from '../../../../constants/actionTypes';
 
-import agent from '../../../agent';
+import agent from '../../../../agent';
 import articlePreviewStyle from './ArticlePreview.module.css';
 
 const FAVORITED_CLASS = 'btn btn-sm btn-primary';
